requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
supabase==1.1.2
